refactor: 同步评论/点赞/收藏相关接口字段并对齐前后端调用 1.统一后端 Article/Comment/Favorite/Like 相关 Controller、Service 及实现层的方法签名与调用参数。 2.同步调整评论与点赞相关 DTO/Entity/VO 字段，保持返回结构与业务层处理一致。 3.前端博客端同步更新评论组件与文章 API 调用，适配后端字段变更，修正评论展示与回复交互中的参数传递。

+++ blog-backend/src/main/java/xyz/kuailemao/service/impl/ArticleServiceImpl.java
@@ -164,7 +164,7 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
     }
 
     @Override
-    public ArticleDetailVO getArticleDetail(Integer id) {
+    public ArticleDetailVO getArticleDetail(Long id) {
         Article article = articleMapper.selectOne(new LambdaQueryWrapper<Article>().eq(Article::getStatus, SQLConst.PUBLIC_ARTICLE).and(i -> i.eq(Article::getId, id)));
         if (StringUtils.isNull(article)) return null;
         // 文章分类
@@ -197,7 +197,7 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
     }
 
     @Override
-    public List<RelatedArticleVO> relatedArticleList(Integer categoryId, Integer articleId) {
+    public List<RelatedArticleVO> relatedArticleList(Long categoryId, Long articleId) {
         // 文章id不等于当前文章id,相关推荐排除自己，5条
         List<Article> articles = articleMapper.selectList(
                 new LambdaQueryWrapper<Article>()
